Ref: util module refactored
